+++ docs/changelog.rst
@@ -1,6 +1,21 @@
 Release History
 ===============
 
+v0.8.0
+------
+
+**Disclaimer**
+
+All releases in the v0.x.x series are subject to breaking changes from one version to another. After the release of v1.0.0, this project will be subject to `semantic versioning <http://semver.org/>`_.
+
+**New Features and Enhancements**
+
+* Added support for deleting entities
+* Added support for interacting with scorecard templates
+* Added support for getting the current user session
+* Added support for uploading files directly to a patient
+* Added support for tracking upload progress
+
 v0.7.0
 ------